fix: switch to eslint and add husky
