+++ assembly/__tests__/characters.ts
@@ -1,0 +1,11 @@
+import { expectMatch, expectNotMatch } from "./utils";
+
+it("single character", () => {
+  expectMatch("a", ["a"]);
+  expectNotMatch("a", ["fish", ""]);
+});
+
+it("concatenation", () => {
+  expectMatch("ab", ["ab"]);
+  expectNotMatch("ab", ["aac", "aa", ""]);
+});
