+++ assembly/nfa/nfa.ts
@@ -8,13 +8,13 @@ import {
   CharacterSetNode,
   CharacterClassNode,
   GroupNode,
-  AssertionNode,
   NodeType,
 } from "../parser/node";
 
 import { Char } from "../char";
 import { Matcher } from "./matcher";
 
+/* eslint @typescript-eslint/no-empty-function: ["error", { "allow": ["constructors", "methods"] }] */
 export class State {
   epsilonTransitions: Array<State> = [];
 
@@ -201,10 +201,11 @@ function automataForNode(expression: Node | null): Automata {
       return Automata.fromMatcher(
         Matcher.fromCharacterClassNode(expression as CharacterClassNode)
       );
-    case NodeType.Group:
+    case NodeType.Group: {
       const c = expression as GroupNode;
       const auto = automataForNode(c.expression);
       return group(auto);
+    }
     case NodeType.Assertion:
       return Automata.fromEpsilon();
     default:
