+++ package.json
@@ -9,13 +9,14 @@
     "jest": "jest __tests__ --testPathIgnorePatterns assembly/*",
     "asp": "asp --verbose",
     "asp:ci": "asp",
-    "prettier:check": "prettier --check .",
-    "prettier:write": "prettier --write .",
+    "prettier:check": "npm run eslint",
+    "prettier:write": "npm run eslint -- --fix ",
     "asbuild:untouched": "asb --target debug",
     "asbuild:optimized": "asb",
     "asbuild": "npm run asbuild:untouched && npm run asbuild:optimized",
     "tsrun": "ts-node ts/index.ts",
-    "benchmark": "node benchmark/benchmark.js"
+    "benchmark": "node benchmark/benchmark.js",
+    "eslint": "eslint --max-warnings 0 --ext ts \"assembly/**/*.ts\""
   },
   "author": "colin.eberhardt@gmail.com",
   "license": "MIT",
@@ -23,16 +24,24 @@
     "@as-pect/cli": "^5.0.1",
     "@assemblyscript/loader": "^0.17.5",
     "@types/node": "^14.14.13",
+    "@typescript-eslint/eslint-plugin": "^4.14.1",
+    "@typescript-eslint/parser": "^4.14.1",
     "asbuild": "0.0.10",
     "assemblyscript": "0.17.5",
     "benchmark": "^2.1.4",
+    "eslint": "^7.18.0",
     "jest": "^26.6.3",
     "jest-summary-reporter": "0.0.2",
-    "prettier": "^2.2.1",
     "text-encoding": "^0.7.0",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3",
-    "visitor-as": "^0.3.2"
+    "visitor-as": "^0.3.2",
+    "husky": "^4.2.5"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run prettier:check || (yarn prettier:fix && exit 1)"
+    }
   },
   "dependencies": {}
 }
